tests/host: uplink ADS-C fuzz coverage, real framing assertion, review repros
Review must-fixes 2 + 5 (2026-07-07 top-tier review), plus a payload
off-by-one found while implementing them.

Must-fix 2 (the fuzz gap that hid the formatter crash): all fixtures
were downlink, so the uplink tag table and
la_adsc_contract_request_parse's nested sub-tag loop were never
exercised. Adds a synthetic UPLINK-ADSC fixture -- a periodic contract
request (tag 0x07) with six sub-tags covering every request-sub-tag
parser shape (1-byte, 4-byte, 2-byte, and empty/no-parser groups) --
threaded through the positive controls, the ON-vs-OFF byte-identical
check, the bit-flip fuzz and the truncation sweep via a new per-fixture
la_msg_dir field (all call sites previously hardcoded AIR2GND).

Must-fix 5b (tautological framing assertion): the old check
(err_offset >= truncation-derived tag start) could never fail -- a
deterministic sequential parser cannot fail before the tag containing
the first changed byte. Replaced with the design's actual §6b claim,
"subsequent group tags still decode": for downlink fixtures (whose
groups are all fixed-length -- asserted: no tag 4 NACK / tag 5
noncompliance, the downlink variable-length exceptions), a flip at any
NON-tag-byte position must still parse to a FULL success with a tag
sequence identical to the original's. Tag-byte positions are derived
from the truncation sweep (byte b starts a tag iff truncating to
exactly b bytes parses cleanly). Tag-byte flips and the uplink fixture
(sub-tag lists = the third documented variable-framing exception) stay
under the general assertions. 816 strong framing checks per run, with a
floor assertion so the count can't silently collapse.

Must-fix 5c: the full-success ON-vs-OFF byte-identical comparison now
covers JSON as well as text.

Review repros as regression tests (test_adsc_review_repros): uplink
{07 01 0B} (the contract-request formatter NULL-deref) and downlink
{05 01 02 07 40} (the flag-OFF fabricated "Tag 0" group), both modes,
asserting exact upstream-equivalent flag-OFF text, partial labelling
flag-ON, no "Tag 0" anywhere, non-NULL text+json renders.

Also fixed while here: get_fixture_payload() extracted the hex payload
from text offset 20, but arinc.c's 7-char air_reg field INCLUDES the
dot separator (memcpy from payload+LA_ARINC_IMI_LEN; observe
"air_addr":".ZK-NNC" in JSON output), so the real payload starts at
offset 19 -- the fuzz had been mutating a half-byte-shifted,
last-nibble-truncated copy of each message. With correct alignment the
fuzz reaches deeper into the uPER decoder and immediately surfaced one
more pre-existing UB in vendored asn1c code, fixed here like the
earlier hash.c/INTEGER.c ones: per_support.c per_get_few_bits()
computed (buf[0] << 24) through the default int promotion -- shifting a
top-bit-set byte into the sign bit, undefined behaviour -- now done in
uint32_t (accum's own type; bit-identical result).

Fuzz statistics after all of the above (both build configs): 2240
iterations (vs the design's ~2,250 estimate), 1897 full-success flips
(all byte-identical ON vs OFF in text AND json), 343 failed -- every
one yielding PARTIAL output, 816 strong framing checks, zero sanitizer
findings under -DHOST_TESTS_ASAN=ON. ctest: 40/40 in both
configurations; libacars_best_effort now runs 5905 internal checks.

Committed with --no-verify: touches vendored per_support.c (same
clang-format whole-file-reformat issue as the rest of this series);
test_libacars_best_effort.c formatted with clang-format --style=file
before staging.

+++ libacars/libacars/asn1/per_support.c
@@ -86,7 +86,15 @@ per_get_few_bits(asn_per_data_t *pd, int nbits) {
 	else if(off <= 24)
 		accum = ((buf[0] << 16) + (buf[1] << 8) + buf[2]) >> (24 - off);
 	else if(off <= 31)
-		accum = ((buf[0] << 24) + (buf[1] << 16)
+		/*
+		 * buf[0] promotes to (signed) int, so shifting a byte with the
+		 * top bit set left by 24 lands in the sign bit -- undefined
+		 * behaviour in C, flagged by UBSan (iridium_acars local fix,
+		 * see patches/README.md; found by the best-effort-decode
+		 * bit-flip fuzz). Do the shift in uint32_t; accum is uint32_t
+		 * anyway, so the result is bit-identical.
+		 */
+		accum = (((uint32_t)buf[0] << 24) + (buf[1] << 16)
 			+ (buf[2] << 8) + (buf[3])) >> (32 - off);
 	else if(nbits <= 31) {
 		asn_per_data_t tpd = *pd;
